Bot/requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
aiohttp-socks==0.8.4
fastapi==0.111.0
databases==0.9.0